Docker update for rocBLAS, HCC, and HIP to 2.1 (#1606)
* Updated the requirements with newer HIP, HCC, rocBLAS, and clang-ocl

* Fixed boost issue.

* Fixed tidy issues across the lib.

* Formatting.

* Update hip and rocblas version.

* Fixed remaining tidy issues from formatting.

* Formatting.

* More tidy issues.

* Float conversion.

* NOLINT on stubborn coversion.

* ktime to float

* Remove unneeded NOLINT

* Removed unneeded semi-colon

* Drive updates.

* Switched to rocBLAS 2.3 RC commit hash.

* Use size_t for fusion op_map.

+++ src/kernel_cache.cpp
@@ -80,7 +80,7 @@ static void AddKernelDumpKernelParams(const std::string& program_name,
         if(q == nullptr)
             break;
         q++;
-        for(int i = 0; i < sizeof(keys) / sizeof(keys[0]); i++)
+        for(unsigned long i = 0; i < sizeof(keys) / sizeof(keys[0]); i++)
         {
             if(strncmp(p, keys[i], strlen(keys[i])) == 0)
             {
