+++ src/pooling.cpp
@@ -109,22 +109,26 @@ PoolingDescriptor::GetForwardOutputDim(const TensorDescriptor& tensorDesc) const
 
     auto output_h = std::max<std::ptrdiff_t>(
         1,
-        std::ptrdiff_t(std::ceil((input_h + 2 * pad_h - window_h) / static_cast<float>(stride_h)) +
+        std::ptrdiff_t(std::ceil(static_cast<float>(input_h + 2 * pad_h - window_h) /
+                                 static_cast<float>(stride_h)) +
                        1));
     auto output_w = std::max<std::ptrdiff_t>(
         1,
-        std::ptrdiff_t(std::ceil((input_w + 2 * pad_w - window_w) / static_cast<float>(stride_w)) +
+        std::ptrdiff_t(std::ceil(static_cast<float>(input_w + 2 * pad_w - window_w) /
+                                 static_cast<float>(stride_w)) +
                        1));
 
     if(_pMode == miopenPaddingSame)
     {
-        output_h = std::ceil(static_cast<double>(input_h) / stride_h);
-        output_w = std::ceil(static_cast<double>(input_w) / stride_w);
+        output_h = std::ceil(static_cast<double>(input_h) / static_cast<double>(stride_h));
+        output_w = std::ceil(static_cast<double>(input_w) / static_cast<double>(stride_w));
     }
     else if(_pMode == miopenPaddingValid)
     {
-        output_h = std::ceil(static_cast<double>(input_h - window_h + 1) / stride_h);
-        output_w = std::ceil(static_cast<double>(input_w - window_w + 1) / stride_w);
+        output_h =
+            std::ceil(static_cast<double>(input_h - window_h + 1) / static_cast<double>(stride_h));
+        output_w =
+            std::ceil(static_cast<double>(input_w - window_w + 1) / static_cast<double>(stride_w));
     }
 
     return std::make_tuple(input_n, input_c, output_h, output_w);
